Moved api routes to seperate file
